requirements.txt
annotated-types==0.7.0
anyio==4.8.0
bitarray==3.0.0
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
ckzg==2.0.1
click==8.1.8
cytoolz==1.0.1
dash==2.18.2
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.1.0
eth-utils==5.2.0
eth_abi==5.2.0
Flask==3.0.3
h11==0.14.0
h2==4.2.0
hexbytes==1.3.0
hpack==4.1.0
httpcore==1.0.7
httpx==0.25.2
hyperframe==6.1.0
idna==3.10
importlib_metadata==8.6.1
iniconfig==2.0.0
itsdangerous==2.2.0
Jinja2==3.1.5
joblib==1.4.2
MarkupSafe==3.0.2
nest-asyncio==1.6.0
numpy>=1.21.0
packaging==24.2
pandas>=1.5.0
parsimonious==0.10.0
pipe==2.2
plotly==5.24.1
pluggy==1.5.0
poly_eip712_structs==0.0.1
py_clob_client==0.20.0
py_order_utils==0.3.2
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pytest==8.3.4
pytest-asyncio==0.25.3
python-dateutil==2.9.0.post0
python-dotenv>=0.19.0
pytz==2025.1
regex==2024.11.6
requests>=2.26.0
retrying==1.3.4
rlp==4.1.0
scikit-learn>=1.0.0
scipy==1.15.1
setuptools==75.8.0
six==1.17.0
sniffio==1.3.1
subgrounds>=1.0.0
tenacity==9.0.0
threadpoolctl==3.5.0
toolz==1.0.0
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
Werkzeug==3.0.6
zipp==3.21.0
